requirements.txt
# Core Framework
flask==3.0.0
python-dotenv==1.0.0

# AI Integration
openai==1.13.3
tiktoken==0.6.0
tenacity==8.2.3

# File Processing
markdownify==0.11.6
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27
chardet==5.2.0
beautifulsoup4==4.12.3

# File System Monitoring
watchdog==4.0.0
aiofiles==23.2.1

# Web & API
werkzeug==3.0.1
flask-uploads==0.2.1
jinja2==3.1.3
flask-cors==4.0.0

# Security & Utils
cryptography==42.0.5
uuid6==2024.1.12
requests==2.31.0

# Data Processing
pathlib2==2.3.7
jsonschema==4.21.1

# Development & Testing
pytest==8.1.1
pytest-flask==1.3.0
black==24.3.0
flake8==7.0.0

# Production
gunicorn==21.2.0
procfile==0.1.0